% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSplines.R
\name{brspl}
\alias{brspl}
\title{Bayesian fitting of shape-restricted regression spline.}
\usage{
brspl(type, x, y, k, nloop)
}
\arguments{
\item{type}{type of the algorithm (and the restriction on the shape)
1 = monotone increasing (quadratic)}

\item{x}{vector of n values of the one-dimensional covariate}

\item{y}{vector of n values of the response variable}

\item{k}{number of interior knots}

\item{nloop}{number of loops in the algorithm}
}
\value{
a list of elements
fit: point fitted values
lower: lower bound of the credible interval
upper: upper bound of the credible interval
coefs: coefficients of the spline regression
}
\description{
This function fit a Bayesian shape-restricted regression spline, using a user-chosen algorithm type.
Code available on Prof. Mary C. Meyer's website (Colorado State University) https://www.stat.colostate.edu/~meyer/bayescode.htm
}
\keyword{CondCop}
