% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSplines.R
\name{concave}
\alias{concave}
\title{Concave function}
\usage{
concave(x, t)
}
\arguments{
\item{x}{value of the covariate}

\item{t}{value of t}
}
\value{
value
}
\description{
Concave function
}
\keyword{CondCop}
