% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP.R
\name{cond_GP}
\alias{cond_GP}
\title{Gaussian process modelling for the Fisher transform of the condtional functionals.}
\usage{
cond_GP(z, nz, y, model = "bgp")
}
\arguments{
\item{z}{values of the covariate}

\item{nz}{number of values in the grid of values in the domain of the covariate. Default at 1000.}

\item{y}{response variable. Typically this is the Fisher transform of the estimated conditional functional.}

\item{model}{Bayesian regression models to be implemented. "blm" stands for linear model,
"bgpllm" stands for GP with jumps to the limiting linear model (LLM),
"bgp" stands for Gaussian process. Default at "bgp"}
}
\value{
a list including
estim: point estimate for the functional of interest phi
q1: quantile of level 0.025 for the functional of interest phi
q2: quantile of level 0.975 for the functional of interest phi
xgrid: a vector of points representing a grid of values on which to approximate the predictive distribution

x.grid grid of ngrid values relative to each row of the matrix rho
}
\description{
This function makes use of the tgp package developed by Robert B. Gramacy (Virginia Polytechnic and State University)
and Dr. Matt A. Taddy (Amazon)
https://cran.r-project.org/web/packages/tgp/tgp.pdf
}
\details{
X=z, XX=x.grid, Z=W_rep
}
\keyword{CondCop}
