% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSplines.R
\name{decconcave}
\alias{decconcave}
\title{Decreasing concave function}
\usage{
decconcave(x, t)
}
\arguments{
\item{x}{value of the covariate}

\item{t}{value of t}
}
\value{
value
}
\description{
Decreasing concave function
}
\keyword{CondCop}
