% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSplines.R
\name{decconvex}
\alias{decconvex}
\title{Decreasing convex function}
\usage{
decconvex(x, t)
}
\arguments{
\item{x}{value of the covariate}

\item{t}{value of t}
}
\value{
value
}
\description{
Decreasing convex function
}
\keyword{CondCop}
