% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{EL}
\alias{EL}
\title{Empirical likelihood}
\usage{
EL(SC)
}
\arguments{
\item{SC}{individual contributions of the estimating function.
It can be either an n x 1 vector (scalar parameter) or an n x p matrix (vector valued parameter)}
}
\value{
w0: weights associated to each unit

conv: convergence of the algorithm (available only for vector valued parameter)

elr: empirical loglikelihood ratio

el: empirical loglikelihood
}
\description{
This function computes the empirical likelihood by minimising a quadratic function.
}
\keyword{CondCop}
