% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ELCOP.cond}
\alias{ELCOP.cond}
\title{Nonparametric Bayesian estimate of the conditional Spearman's rho through the empirical likelihood, with
a linearised model.}
\usage{
ELCOP.cond(w, x, p, nsim = 10000, mu = NULL, Sigma = NULL)
}
\arguments{
\item{w}{input data, i.e. Fisher transform of the estimated Spearman's rho.}

\item{x}{vector of n values of the one-dimensional covariate}

\item{p}{order of the approximation in the Taylor expansion.}

\item{nsim}{number of Monte Carlo simulations. Default at 10,000.}

\item{mu}{prior mean of the coefficients. If NULL, the mean will be a vector of zeros of length p. Default at NULL}

\item{Sigma}{prior covariance matrix of the coefficients of the linearised model. If NULL, it is fixed at a diagonal
matrix with elements on the diagonal equal to 10. Default at NULL.}
}
\value{
beta: a matrix of dimension nsim x p with a sample of coefficients parameters

omega: a vector of nsim weights representing the empirical likelihood for the coefficients in the matrix beta.
}
\description{
This function provides a nonparametric approximation of the posterior distribution of the conditional Spearman's rho.
This function produces estimates computing the likelihood for the linearised model of the conditional Spearman's rho,
obtained through a Taylor expansion.
}
\keyword{CondCop}
