% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ELCOP.taucov}
\alias{ELCOP.taucov}
\title{Nonparametric Bayesian estimate of the conditional Kendall's tau through the empirical likelihood.}
\usage{
ELCOP.taucov(
  uu,
  x,
  meth,
  nsim = 10000,
  bdw,
  kern = "gauss",
  alpha.gp = 0.1,
  ngrid = 30
)
}
\arguments{
\item{uu}{n x 2 matrix of pseudo-observations. n is the sample size.}

\item{x}{vector of n values of the one-dimensional covariate}

\item{meth}{weights to use in the computation of the conditional copula. Two alternatives: "NW" for
Nadaraya-Watson weights, "LL" for local-linear weights}

\item{nsim}{number of Monte Carlo simulations. Default at 10,000.}

\item{bdw}{bandwidth}

\item{kern}{kernel function to use to compute the local-linear weights. Two alternatives: either "gauss" for Gaussian
kernel or "t" for triweight kernel Default: "gauss".}

\item{ngrid}{number of points in the prediction grid for the covariate. Default at 30.}
}
\value{
tau matrix ngrid x 3 with median and posterior credible intervals of level 0.90 for the conditional
Kendall's tau.

x.grid grid of ngrid values relative to each row of the matrix tau
}
\description{
This function provides a nonparametric approximation of the posterior distribution of the conditional
Kendall's tau.
This function produces estimates using the inconsistent estimator of the conditional copula.
}
\keyword{CondCop}
