% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fish_transf}
\alias{fish_transf}
\title{Fisher transform}
\usage{
fish_transf(y)
}
\arguments{
\item{y}{vector to compute the Fisher transform}
}
\value{
a vector of the same length of y with the values of the Fisher transform
}
\description{
This function computes the Fisher transform for a vector y.
}
\keyword{CondCop}
