% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gaussweight}
\alias{gaussweight}
\title{Gaussian kernel function}
\usage{
gaussweight(x)
}
\arguments{
\item{x}{scalar or vector on which to compute the Gaussian function}
}
\value{
value of the Gaussian function
}
\description{
This function computes the Gaussian kernel function of an input vector
}
\keyword{CondCop}
