% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSplines.R
\name{incconcave}
\alias{incconcave}
\title{Increasing concave function}
\usage{
incconcave(x, t)
}
\arguments{
\item{x}{value of the covariate}

\item{t}{value of t}
}
\value{
value
}
\description{
Increasing concave function
}
\keyword{CondCop}
