% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSplines.R
\name{incconvex}
\alias{incconvex}
\title{Increasing convex function}
\usage{
incconvex(x, t)
}
\arguments{
\item{x}{value of the covariate}

\item{t}{value of t}
}
\value{
value
}
\description{
Increasing convex function
}
\keyword{CondCop}
