% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LL.weights}
\alias{LL.weights}
\title{Local-linear weights}
\usage{
LL.weights(x.vec, x, bdw, kern = "gauss")
}
\arguments{
\item{x.vec}{vector on which to compute the local-linear weights}

\item{x}{scalar: this is the point with respect to which to compute the local-linear weights for each of the
points of x.vec}

\item{bdw}{bandwidth}

\item{kern}{kernel function to use to compute the local-linear weights. Two alternatives: either "gauss" for Gaussian
kernel or "t" for triweight kernel Default: "gauss".}
}
\value{
vector of weights
}
\description{
This function computes the local-linear function weights
}
\keyword{CondCop}
