% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSplines.R
\name{mondecr}
\alias{mondecr}
\title{Monotone decreasing function}
\usage{
mondecr(x, t)
}
\arguments{
\item{x}{value of the covariate}

\item{t}{value of t}
}
\value{
value
}
\description{
Monotone decreasing function
}
\keyword{CondCop}
