% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSplines.R
\name{monincr}
\alias{monincr}
\title{Monotone increasing function}
\usage{
monincr(x, t)
}
\arguments{
\item{x}{value of the covariate}

\item{t}{value of t}
}
\value{
value
}
\description{
Monotone increasing function
}
\keyword{CondCop}
