% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{NW.weights}
\alias{NW.weights}
\title{Nadaraya-Watson weights}
\usage{
NW.weights(x.vec, x, bdw, kern = "gauss")
}
\arguments{
\item{x.vec}{vector on which to compute the Nadaraya-Watson weights}

\item{x}{scalar: this is the point with respect to which to compute the Nadaraya-Watson weights for each of the points of x.vec}

\item{bdw}{bandwidth}

\item{kern}{kernel function to use to compute the Nadaraya-Watson weights. Two alternatives: either "gauss" for Gaussian
kernel or "t" for triweight kernel Default: "gauss".}
}
\value{
vector of weights
}
\description{
This function computes the Nadaraya-Watson function weights
}
\keyword{CondCop}
