% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{NW.weights_multi}
\alias{NW.weights_multi}
\title{Multivariate Nadaraya-Watson weights}
\usage{
NW.weights_multi(x.mat, x.vec, bdw, kern = "gauss")
}
\arguments{
\item{x.mat}{matrix of value on which to compute the Nadaraya-Watson weights}

\item{x.vec}{vector: this is the point with respect to which to compute the Nadaraya-Watson weights for each of the points of x.vec}

\item{bdw}{bandwidth}

\item{kern}{kernel function to use to compute the Nadaraya-Watson weights. Two alternatives: either "gauss" for Gaussian
kernel or "t" for triweight kernel Default: "gauss".}
}
\value{
vector of weights
}
\description{
This function computes the multivariate Nadaraya-Watson function weights
}
\keyword{CondCop}
