% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Owen}
\alias{Owen}
\title{Quadratic function for the empirical likelihood}
\usage{
Owen(x, thre, gradient)
}
\arguments{
\item{x}{observations}

\item{thre}{threshold}

\item{gradient}{gradient}
}
\value{
A vector representing the pseudo-logaritm of the values to minimise in the empirical likelihood
}
\description{
This function implements the quadratic function in Owen (2001) (Section 3.14).
This is the function to optimise to obtain the empirical likelihood.
}
\keyword{CondCop}
