% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSplines.R
\name{postgen}
\alias{postgen}
\title{Posterior generation for the beta coefficients in shape-restricted splines.}
\usage{
postgen(a, b, c)
}
\arguments{
\item{a}{parameter a in the regression spline model}

\item{b}{parameter b in the regression spline model}

\item{c}{parameter c in the regression spline model}
}
\value{
transformed value for beta
}
\description{
Posterior generation for the beta coefficients in shape-restricted splines.
}
\keyword{CondCop}
