% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{rho.cond}
\alias{rho.cond}
\title{Conditional Spearman's rho.}
\usage{
rho.cond(uu, meth = "NW", x.vec, xgrid, bdw, kern = "gauss")
}
\arguments{
\item{uu}{n x 2 matrix of pseudo-observations. n is the sample size.}

\item{meth}{weights to use in the computation of the conditional copula. Two alternatives: "NW" for
Nadaraya-Watson weights, "LL" for local-linear weights}

\item{x.vec}{vector of n values of the one-dimensional covariate}

\item{xgrid}{vector representing a grid of values with respect to which to compute the weights}

\item{bdw}{bandwidth}

\item{kern}{kernel function to use to compute the local-linear weights. Two alternatives: either "gauss" for Gaussian
kernel or "t" for triweight kernel Default: "gauss".}
}
\value{
nonparametric point estimate of the Spearman's rho.
}
\description{
This function provides a nonparametric point estimate for the conditional Spearman'rho
}
\keyword{CondCop}
