% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionals.R
\name{tau.cond_multi}
\alias{tau.cond_multi}
\title{Conditional Kendall's tau for multivariate covariates.}
\usage{
tau.cond_multi(uu, x.mat, x1.grid, x2.grid, bdw, kern = "gauss")
}
\arguments{
\item{uu}{n x 2 matrix of pseudo-observations. n is the sample size.}

\item{x.mat}{nx2 matrix of values of a 2-dimensional covariate}

\item{x1.grid}{vector representing a grid of values for covariate 1}

\item{x2.grid}{vector representing a grid of values for covariate 2}

\item{bdw}{bandwidth}

\item{kern}{kernel function to use to compute the NW weights. Two alternatives: either "gauss" for Gaussian
kernel or "t" for triweight kernel Default: "gauss".}

\item{meth}{weights to use in the computation of the conditional copula. Two alternatives: "NW" for
Nadaraya-Watson weights, "LL" for local-linear weights}
}
\value{
matrix with nonparametric point estimate of the Kendall's tau for each combination of the values
in x1.grid and x2.grid
}
\description{
This function provides a nonparametric point estimate for the conditional Kendall's tau
}
\keyword{CondCop}
