% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{triweight}
\alias{triweight}
\title{Triweight kernel function}
\usage{
triweight(x)
}
\arguments{
\item{x}{scalar or vector on which to compute the triweight function}
}
\value{
value of the triweight function
}
\description{
This function computes the triweight kernel function of an input vector
}
\keyword{CondCop}
